# docs/api.yml
openapi: 3.0.3
info:
  title: Docbox API
  version: 1.0.0
  description: API specification for docbox
paths:
  /admin/tenant:
    post:
      summary: Create a tenant
      description: >
        Administrative endpoint creates a new tenant, performs tenant setup such as:
        - Creating the tenant in the root database
        - Connect to the tenant database and setup the database tables 
        - Create the tenant S3 bucket and attach bucket queue events
        - Create the tenant opensearch index
      operationId: create_tenant
      parameters:
        - $ref: "#/components/parameters/x-tenant-env"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  allOf:
                    - $ref: "#/components/schemas/TenantId"
                    - description: Unique ID to give the tenant
              required:
                - id
      responses:
        "201":
          description: Tenant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    allOf:
                      - $ref: "#/components/schemas/TenantId"
                      - description: ID of the created tenant (Same as the provided ID)
        "400":
          $ref: "#/components/responses/BadRequest"

        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/tenant/{tenant_id}:
    get:
      summary: Get a tenant
      operationId: get_tenant
      description: >
        Administrative endpoint to request the details of a tenant. Provides the
        details of the tenant itself along with a list of all the document boxes 
        within the tenant
      parameters:
        - $ref: "#/components/parameters/x-tenant-env"
        - name: tenant_id
          in: path
          description: "ID Of the tenant to get"
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant existed
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    allOf:
                      - $ref: "#/components/schemas/Tenant"
                      - description: Configuration for the requested tenant
                  document_boxes:
                    type: array
                    description: List of all document boxes available in the tenant
                    items:
                      $ref: "#/components/schemas/DocumentBox"
        "404":
          $ref: "#/components/responses/NotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"
  /admin/search:
    post:
      summary: Search within a tenant
      operationId: search_tenant
      description: >
        Administrative endpoint to search within a set of document boxes in a provided
        tenant. This is an administrative endpoint because user authentication cannot 
        be enforced on this endpoint because its not apart of a specific document box
        and thus authentication should be performed on the proxy server and performed
        on behalf of the user based on scopes they're allowed access to
      parameters:
        - $ref: "#/components/parameters/x-tenant-env"
        - $ref: "#/components/parameters/x-tenant-id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  properties:
                    scopes:
                      type: array
                      description: >
                        Document box scopes to search within, if none are provided empty
                        search results will be returned
                      items:
                        $ref: "#/components/schemas/DocumentBoxScope"
                  required:
                    - scopes
                - $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_hits:
                    type: number
                    description: Total number of hits from the query
                  results:
                    type: array
                    description: Search result items
                    items:
                      $ref: "#/components/schemas/SearchResultItemV2WithScope"

        "404":
          $ref: "#/components/responses/NotFound"

        "400":
          $ref: "#/components/responses/BadRequest"

        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /box/{scope}/folder/{folder_id}:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
      - $ref: "#/components/parameters/folder_id"
    get:
      summary: Get a folder
      operationId: get_folder
      description: >
        Request a specific folder, responds with the folder itself as well as the 
        children contained in the folder (Depth of 1, not recursive)
      responses:
        "200":
          description: Folder exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderResponse"
        "404":
          description: Folder not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update a folder
      operationId: update_folder
      description: >
        Update a folder, allows changing the folder name and parent folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  nullable: true
                  minLength: 1
                folder_id:
                  $ref: "#/components/schemas/FolderId"
                  nullable: true
      responses:
        "200":
          description: Folder updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Folder not found or target folder could not be found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a folder
      operationId: delete_folder
      description: >
        Deletes a folder and everything contained within the folder
      responses:
        "204":
          description: Folder deleted successfully
        "404":
          description: Folder not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    scope:
      name: scope
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/DocumentBoxScope"
    folder_id:
      name: folder_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/FolderId"
...
    x-tenant-id:
      name: x-tenant-id
      in: header
      description: ID of the tenant the request is for
      required: true
      schema:
        $ref: "#/components/schemas/TenantId"
    x-tenant-env:
      name: x-tenant-env
      in: header
      description: Environment to operate withins
      required: true
      schema:
        type: string
    x-user-id:
      name: x-user-id
      in: header
      description: ID of the user performing the operation (If performed by a user)
      required: false
      schema:
        $ref: "#/components/schemas/UserId"
    x-user-name:
      name: x-user-id
      in: header
      description: Name of the user performing the operation (If performed by a user)
      required: false
      schema:
        type: string
    x-user-image-id:
      name: x-user-image-id
      in: header
      description: Image ID of the user performing the operation (If performed by a user)
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request, response body contains the error message
      content: text/plain
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Tenant:
      type: object
      description: Configuration details about a tenant
      properties:
        id:
          $ref: "#/components/schemas/TenantId"
        db_name:
          type: string
          description: Name of the database the tenant uses
        db_secret_name:
          type: string
          description: Name of the AWS secret containing the database credentials
        s3_name:
          type: string
          description: Name of the S3 bucket storing the tenant data
        os_index_name:
          type: string
          description: Name of te opensearch index containing the tenant data
        env:
          type: string
          description: >
            Environment the tenant is within. Keyed with ID to allow the same ID to
            exist within multiple different environments. Common values:
            - "Development"
            - "Production"
      required:
        - id
        - db_name
        - db_secret_name
        - s3_name
        - os_index_name
        - env
    DocumentBox:
      type: object
      description: Basic details about a document box
      properties:
        scope:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - scope
        - created_at
    SearchRequest:
      type: object
      description: >
        Request for searching
      properties:
        version:
          type: string
          description: Version for the response format
          enum: [V1, V2]
          default: V1
        query:
          type: string
          description: Search query
          minLength: 1
        neural:
          type: boolean
          description: >
            Enable searching with AI, not currently enabled on production
        mime:
          type: string
          description: >
            Filter to searching files with a specific mime type
        include_name:
          type: boolean
          description: >
            Whether to include the name of the item when searching with the query
        include_content:
          type: boolean
          description: >
            Whether to include the contents when searching with the query
        max_fragment_size:
          type: number
          maximum: 150
          default: 150
          description: >
            Maximum size of text fragments to highlight, only works when "include_content"
            is specified
        max_fragments:
          type: number
          maximum: 100
          default: 3
          description: >
            Maximum number of text fragments to extract per page of the file
        created_at:
          allOf:
            - $ref: "#/components/schemas/SearchRange"
            - description: Filter to a specific creation date range
        modified:
          allOf:
            - $ref: "#/components/schemas/SearchRange"
            - description: Filter to a specific modified date range
        created_by:
          allOf:
            - $ref: "#/components/schemas/UserId"
            - description: User ID for a user to filter creation by
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: >
                ID of a folder to filter to the contents of searches within the folder itself
                as well as within all nested folders
        item_id:
          allOf:
            - $ref: "#/components/schemas/ItemId"
            - description: >
                Filter to searching over a specific item only
        size:
          type: number
          description: Number of items to include in the response
          default: 50
        offset:
          type: number
          description: Offset to start querying from
          default: 0
        max_pages:
          type: number
          description: Maximum pages to include in the response for file content
          maximum: 100
          default: 3
        pages_offset:
          type: number
          description: Offset for results within the files pages query
          default: 0
    SearchRange:
      type: object
      description: >
        Range of dates for a search query
      properties:
        start:
          type: string
          format: date-time
          description: Start portion of a date range (Omit to include any dates before the end date)
          nullable: true
        end:
          type: string
          format: date-time
          description: End portion of a date range (Omit to include any dates after the start date)
          nullable: true
    UserId:
      type: string
      description: ID of a user
    TenantId:
      type: string
      format: uuid
      description: ID of a tenant
    ItemId:
      oneOf:
        - $ref: "#/components/schemas/FileId"
        - $ref: "#/components/schemas/FolderId"
        - $ref: "#/components/schemas/LinkId"
...
    FileId:
      type: string
      format: uuid
      description: ID of a file
    FolderId:
      type: string
      format: uuid
      description: ID of a folder
    LinkId:
      type: string
      format: uuid
      description: ID of a link
    DocumentBoxScope:
      type: string
      description: Unique scope for a document box
      minLength: 1
      example: "user:a71502fe-c025-4a06-ae5d-3131535a8237:uploaded"
...
    SearchResultItemV2WithScope:
      description: >
        Search result with the document box scope its within for global search results
        where the specific box must be identified
      allOf:
        - $ref: "#/components/schemas/SearchResultItemV2"
        - type: object
          properties:
            scope:
              $ref: "#/components/schemas/DocumentBoxScope"
    SearchResultItemV2:
      description: >
        Item from a search result (File/Link/Folder) with additional search related information.

        Version 2 for the improved page results and identifying page numbers and additional search
        location information intended for global app search across many document boxes
      allOf:
        - type: object
          properties:
            score:
              type: number
              format: float
              description: Score ranking for the search result
            path:
              type: array
              description: Folder path to the item
              items:
                $ref: "#/components/schemas/FolderPathSegment"
            page_matches:
              type: array
              description: Matches within pages of the item (For files)
              items:
                $ref: "#/components/schemas/PageResult"
            total_hits:
              type: number
              description: Number of matches in the search results query
            name_match:
              type: boolean
              description: Whether the name matches the search query
            content_match:
              type: boolean
              description: Whether the content matches the search query
          required:
            - score
            - path
            - page_matches
            - total_hits
            - name_match
            - content_match
        - $ref: "#/components/schemas/SearchResultData"
    PageResult:
      type: object
      properties:
        page:
          type: number
        matches:
          type: array
          items:
            type: string
    SearchResultData:
      type: object
      description: >
        Search result item (File/Link/Folder), discriminated union based on the "type" property
      discriminator:
        propertyName: type
        mapping:
          File: "#/components/schemas/FileWithExtra"
          Folder: "#/components/schemas/FolderWithExtra"
          Link: "#/components/schemas/LinkWithExtra"
      oneOf:
        - allOf:
            - $ref: "#/components/schemas/FileWithExtra"
            - type: object
              properties:
                type:
                  type: string
                  enum: [File]
        - allOf:
            - $ref: "#/components/schemas/FolderWithExtra"
            - type: object
              properties:
                type:
                  type: string
                  enum: [Folder]
        - allOf:
            - $ref: "#/components/schemas/LinkWithExtra"
            - type: object
              properties:
                type:
                  type: string
                  enum: [Link]
    FolderResponse:
      type: object
      properties:
        folder:
          $ref: "#/components/schemas/FolderWithExtra"
        children:
          allOf:
            - $ref: "#/components/schemas/ResolvedFolderWithExtra"
            - description: >
                All the files, links, folders within the folder along with the path
                to the folder itself
    FolderWithExtra:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/FolderId"
        name:
          type: string
          description: User provided name for the file
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - nullable: true
              description: >
                ID of the parent folder, this is null when the folder is the root 
                folder of a document box
        created_at:
          type: string
          format: date-time
          description: When the folder was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the folder (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the folder has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - document_box
        - created_at
    FileWithExtra:
      type: object
      description: >
        File with additional information about the most recent edit and user information
      properties:
        id:
          $ref: "#/components/schemas/FileId"
        name:
          type: string
          description: Unique ID of the file
        mime:
          type: string
          description: MIME / content type of the file
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder the file is within
        hash:
          type: string
          description: SHA256 hash of the file contents
        size:
          type: number
          description: File size in bytes
        encrypted:
          type: boolean
          description: Whether the file has been detected as encrypted during processing
        created_at:
          type: string
          format: date-time
          description: When the file was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the file (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the file has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - mime
        - folder_id
        - hash
        - encrypted
        - created_at
    LinkWithExtra:
      type: object
      description: >
        Link with additional information about the most recent edit and user information
      properties:
        id:
          $ref: "#/components/schemas/LinkId"
        name:
          type: string
          description: User provided name of the link
        value:
          type: string
          description: The value of the link
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder this link is within
        created_at:
          type: string
          format: date-time
          description: When the link was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the link (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the link has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - value
        - folder_id
        - created_at
...
    User:
      type: object
      description: >
        Snapshot of the external user who performed the action,
        the details for the user are updated every time they
        perform an action
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/UserId"
            - type: string
              description: External ID of the user
        name:
          type: string
          nullable: true
          description: External name of the user if known
        image_id:
          type: string
          nullable: true
          description: External image ID of the user if known
      required:
        - id
...
    ResolvedFolderWithExtra:
      type: object
      description: Contents of a folder and the resolved path to get to the folder
      properties:
        path:
          type: array
          description: Path to the requested folder
          items:
            $ref: "#/components/schemas/FolderPathSegment"
        folders:
          type: array
          description: Folders within the folder
          items:
            $ref: "#/components/schemas/FolderWithExtra"
        files:
          type: array
          description: Files within the folder
          items:
            $ref: "#/components/schemas/FileWithExtra"
        links:
          type: array
          description: Links within the folder
          items:
            $ref: "#/components/schemas/LinkWithExtra"
    FolderPathSegment:
      type: object
      description: Segment of the path to a folder
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder from the segment part
        name:
          type: string
          description: Name of the folder at the segment
    ErrorResponse:
      type: object
      properties:
        reason:
          type: string
          description: Descriptive message for the cause of the error
        backtrace:
          type: string
          description: >
            Server backtrace of the error, only present in local debugging disabled
            for production releases